fix(tailwind): explicitly declare form color option classes
Using string templating breaks detection by tailwind if the color classes
are not used somewhere else in the project.

According to [Tailwind docs](https://tailwindcss.com/docs/detecting-classes-in-source-files#dynamic-class-names)
class names should always exist in full.

This is a bit more verbose and repetitive, but it ensures minimal projects work as expected.